Refactoring code: change the project architecture
