+++ package.json
@@ -2,10 +2,10 @@
   "name": "tg_mini_app_server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "start": "nodemon index.js",
-    "build": "node index.js"
+    "start": "nodemon app.js",
+    "build": "node app.js"
   },
   "author": "",
   "license": "ISC",
